kill cicular deps by moving scope runtimes to a new scope.ts file
